se retorno el tipo de dato float para la propiedad amount en el kardex de materia prima

+++ src/inventoryrawmaterial/entities/inventoryrawmaterial.entity.ts
@@ -8,7 +8,7 @@ export class Inventoryrawmaterial {
   @PrimaryGeneratedColumn('uuid')
   id: string;
 
-  @Column('int',{
+  @Column('float',{
     default: 0
   })
   amount: number;
